[del] binary search code and mods

+++ Library/int128.h
@@ -2,7 +2,7 @@ std::istream& operator>>(std::istream &input, __int128_t &val){
     std::string str_input;
     input >> str_input;
     val = 0;
-    int str_size = str_input.size();
+    int str_size = (int)str_input.size();
     bool neg = str_input[0] == '-';
     for(int i = (neg ? 1 : 0); i < str_size; ++i){
         val = val*10 + (str_input[i] - 48);
